feat(qft): add reusable Quantum Fourier Transform subroutine
